requirements.txt
fastapi==0.104.1
python-dotenv==0.19.0
uvicorn==0.24.0
gunicorn==21.2.0
SQLAlchemy==2.0.38
orjson==3.9.10
asyncpg>=0.28.0
psycopg2-binary==2.9.1
aiosqlite==0.21.0
python-dateutil==2.9.0